+++ examples/example.py
@@ -1,0 +1,1 @@
+# ToDo: put an example of using the code here
