+++ src/antotate/__init__.py
@@ -1,0 +1,4 @@
+import sys, os
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+
+from .antotate import *
